+++ src/infrastructure/applications/rest/config.ts
@@ -17,6 +17,9 @@ export default {
             exposeSentryErrorResponse: normalizeBoolean(process.env.APP_ENVIRONMENT === 'dev' || process.env.SENTRY_EXPOSE_ERROR_RESPONSE),
         },
         isListenLocalhost: normalizeBoolean(process.env.APP_LISTEN_LOCALHOST),
+        newPermissionsCheck: {
+            enabled: normalizeBoolean(process.env.APP_CHECK_NEW_PERMISSIONS),
+        },
     }),
     module: (config: IRestAppModuleConfig) => ({
         ...baseConfig.module(config),